fix checkout page issue

+++ osd/cart/views.py
@@ -79,8 +79,8 @@ def check_out(request, Pickup=False, surcharge=False, overallTotal=0, costTotal=
         template = "checkout1.html"
     else:
         template = "checkout.html"
-    if request.GET.get('surcharge'):
-        surcharge = True
+        if request.GET.get('surcharge'):
+            surcharge = True
     try:
         cart = Cart.objects.get(cart_id=_cart_id(request))
         cart_items = CartItem.objects.filter(cart=cart, active=True)
@@ -128,26 +128,26 @@ def check_out(request, Pickup=False, surcharge=False, overallTotal=0, costTotal=
                 order_item.delete()
             request.session['order_id'] = current_order.order_id
             return redirect('order:thanks')
-    else:
-            try:
-                current_order = Order.objects.create(
-                    total = overallTotal
+    elif request.method != 'GET':
+        try:
+            current_order = Order.objects.create(
+                total = overallTotal
+            )
+            for order_item in cart_items:
+                order_detail = OrderDetail.objects.create(
+                    product = order_item.product.name,
+                    quantity = order_item.quantity,
+                    price = order_item.product.price,
+                    order = current_order
                 )
-                for order_item in cart_items:
-                    order_detail = OrderDetail.objects.create(
-                        product = order_item.product.name,
-                        quantity = order_item.quantity,
-                        price = order_item.product.price,
-                        order = current_order
-                    )
-                    order_detail.save()
-                    # order stock reduced on order save
-                    products = Product.objects.get(id=order_item.product.id)
-                    products.stock = int(order_item.product.stock - order_item.quantity)
-                    products.save()
-                    order_item.delete()
-                request.session['order_id'] = current_order.order_id
-                return redirect('order:thanks')
-            except ObjectDoesNotExist:
-                pass
+                order_detail.save()
+                # order stock reduced on order save
+                products = Product.objects.get(id=order_item.product.id)
+                products.stock = int(order_item.product.stock - order_item.quantity)
+                products.save()
+                order_item.delete()
+            request.session['order_id'] = current_order.order_id
+            return redirect('order:thanks')
+        except ObjectDoesNotExist:
+            pass
     return render(request, template, dict(surcharge=surcharge, Pickup=Pickup,form = form,cart_items = cart_items, overallTotal=overallTotal, costTotal = costTotal, weightTotal = weightTotal, counter = counter, deliveryCost1=deliveryCost1, deliveryCost2 = deliveryCost2))
